Add goals index to the administration
Note we're defining a `Goal` class in the goals controller so it's
easier to use. Since we're in the `Admin::SDG` module, using `SDG::Goal`
would try to load `Admin::SDG::Goal` and so we need to use
`::SDG::Goal`. However, doing so all the time is repetitive. That's why
in the past we named a controller `Admin::BudgetInvestmentsController`
instead of naming it `Admin::Budget::InvestmentsController`, which IMHO
makes more sense and has less disadvantages but forces us to use
`::Budget` instead of `Budget`.