Disable Better HTML checks
Since we included the pronto-erb_lint gem, we were getting errors in
development where our ERB does not follow the conventions Better HTML
expects.

We only use Better HTML because ERB Lint depends on it, and right now we
are not ready to follow its conventions, so we're disabling it.