Fix flaky empty annotation selection in specs
We were selecting a word using a double click. However, the test failed
sometimes when styles such as font family, font size or the element's
width changed, since the double click takes place in the middle of the
element and it's possible that the middle of the element contains a
space between two words, meaning no word would be selected.

Without these changes, the failure in the "Create" test can be
reproduced by changing the `$base-font-size` variable to `2rem` or
changing the default font family to Roboto.

So we're emulating a triple click on the element to select a whole line
instead. First we click on the element to get the cursor there, and then
we perform a Selenium action three times in a row to emulate the triple
click. Using just `3.times { element.click }` does not work as needed.