Configure EC2 instances to stream logs to AWS CloudWatch Logs
EC2 instances are the machines that run the application code; each
instance runs the puma application server and the delayed_job
process; both of them generate a log that we want to store in a
centralized place.

As Elastic Bean Stalk provides autoscaling, the amount of EC2
instances could vary over time. With this script when the autoscaling
group adds a new instance it will start to stream its logs to
AWS CloudWatch Logs, so you will not loose terminated instances logs
when the environment scales down the EC2 instances.

Original script extracted from here:
https://github.com/awsdocs/elastic-beanstalk-samples/blob/master/configuration-files/aws-provided/instance-configuration/logs-streamtocloudwatch-linux.config