Group links to create budgets with the heading
We're going to add help content here and having a floating button would
make the button disappear when creating content for mobile phones.

I'm not sure having the link in the <header> tag is semantically
correct, but is consistent to what we did to create local targets in the
SDG content section.

Note we're changing the style of the link to create local targets so
it's consistent with the link to create budgets/groups/headings.