Fix missing alt attribute in phases images
An image without an alt text is invalid HTML and is confusing for screen
reader users.

This is just a quick patch which partially solves this problem. The
image doesn't necessarily need to be a decorative one, so administrators
should have the option to provide an alternative text for the images.