Use a custom dropdown widget to switch languages
As mentioned in commit 5214d89c8, using the `change` event of a `select`
field to automatically change location causes accessibility issues to
keyboard users.

So we're implementing our own widget but with a better keyboard
experience.

This is generally a bad idea; the first rule of ARIA is "don't use
ARIA", and we're using a mix of Foundation's code and our custom code to
provide the same functionality as native select fields do.

However, it's better than submitting the form on change.

An alternative here would be to add a submit button. However, with a
submit button users would have to first select an option and then click
the button, which is an extra step. It wouldn't be that big of a deal,
but consider the scenario where users can't read the characters used in
the current language. People will see a button with unfamiliar symbols
and will have to guess that clicking the button will allow them to
change the language they've selected, which could be tricky. As far as
they know, they could be pressing a button to download an app in the
selected language.

So it's important to make it possible for users to switch to their
preferred language just by clicking on its name.

The widget we're adding could be tricky for screen reader users. When
their screen reader announces something in a language they don't
understand, will they guess this is a widget to change the language? Not
all screen readers support the roles we're using and so they might not
announce the control is a menu. In this case, a `<select>` tag at least
is universally supported and it'll be easier for users to realize
they're inside a form to change the language.

I'm also not sure whether the "menuitem" role Foundation provides is the
right role here. I didn't find an obviously better option.

Given these potential issues, we might need to improve the widget in the
future.

Note we're adding a `keydown` event handler to open the menu when the
down arrow key is pressed because most users will expect something to
happen when pressing the down arrow key on something which looks like a
`<select>` field. By default, Foundation opens the menu with the right
arrow key, so we're triggering the right arrow key event when the down
arrow key is pressed.

We're also adding support for typing a letter and going to the language
starting with that letter, which is what some users expect on `<select>`
fields. This solution is inspired by the WAI ARIA collapsible listbox
example [1].

One last note; there's a kind of complex calculation in the padding for
the option links:

```
padding: $line-height / 2 calc(1em + 20px) $line-height / 2 calc(1em + #{rem-calc(4)});
```

The right padding adds 20px to 1em in order to take scrollbars into
account (otherwise the scrollbar could be on top of the text). And the
left padding adds some space for the icon we use to mark the selected
option.

[1] https://www.w3.org/TR/wai-aria-practices-1.1/examples/listbox/listbox-collapsible.html